seperate impl for obtaining application entries
