+++ src/platform/mac.rs
@@ -63,12 +63,12 @@ pub fn get_application_data(path: &Path) -> Option<AppData> {
     })
 }
 
-pub fn get_applications_folders() -> Vec<PathBuf> {
+pub fn get_application_files() -> Vec<PathBuf> {
     let user_dir = PathBuf::from("/Users")
         .join(whoami::username())
         .join("Applications");
 
-    return vec![
+    let application_folders = vec![
         PathBuf::from("/Applications"),
         PathBuf::from("/Applications/Chromium Apps"),
         PathBuf::from("/System/Applications/Utilities"),
@@ -82,6 +82,21 @@ pub fn get_applications_folders() -> Vec<PathBuf> {
         user_dir.clone().join("Chrome Apps.localized"),
         user_dir.clone().join("Brave Apps.localized"),
     ];
+
+    let files = Vec::new();
+
+    for applications_folder in applications_folders {
+        let dir = applications_folder.read_dir();
+        if dir.is_err() {
+            continue;
+        }
+        for entry in dir.unwrap().flatten() {
+            let path = entry.path();
+            files.push(path);
+        }
+    }
+
+    return files;
 }
 
 pub fn get_frontmost_application_data() -> Option<AppData> {
